fix($animate): clear the GCS cache even when no animation is detected
$animate aims to speed up redundant calls to getComputedStyle to prevent
unessarry reflows in the event that similar elements are being animated
at the same time. A hidden bug was causing the GCS cache not to be reset
if an CSS-based animation was not found. This patch ensures that the
GCS lookup cache is always cleared after one animation reflow has
passed.

Closes #8813